LaTeX and divided versions of files.
File VM15042020L has identical code as before, but it is broken down into clearer cells and has some surrounding text, as we need to explain our steps. 
File VM15042020LnoNs is the above file, but then I have removed the redundant Ns from functions update and opara at the start and throughout where necessary. Please merge this to the official VM file when in agreement whether this is/not the correct method, if required.